feat: moved the cache to the DAL layer
