+++ apps/api/src/.env.development
@@ -14,6 +14,9 @@ REDIS_HOST=localhost
 REDIS_PREFIX=
 REDIS_DB_INDEX=2
 
+CACHE_HOST="localhost"
+CACHE_PORT="6379"
+
 NEW_RELIC_ENABLED=true
 NEW_RELIC_APP_NAME="[DEV] - api"
