+++ apps/api/src/.env.production
@@ -14,6 +14,9 @@ REDIS_PORT=6379
 REDIS_PREFIX=
 REDIS_DB_INDEX=2
 
+CACHE_HOST="localhost"
+CACHE_PORT="6379"
+
 NEW_RELIC_ENABLED=true
 NEW_RELIC_APP_NAME="[PROD] - api"
