+++ apps/api/src/.env.test
@@ -10,6 +10,10 @@ REDIS_PORT=6379
 REDIS_HOST=localhost
 REDIS_PREFIX=
 REDIS_DB_INDEX=1
+
+CACHE_HOST="localhost"
+CACHE_PORT="6379"
+
 SYNC_PATH=http://localhost:3001
 API_ROOT_URL=http://localhost:3000
 DISABLE_USER_REGISTRATION=false
