+++ apps/api/src/.example.env
@@ -11,6 +11,9 @@ REDIS_PREFIX=
 REDIS_HOST=localhost
 REDIS_DB_INDEX=2
 
+CACHE_HOST="localhost"
+CACHE_PORT="6379"
+
 JWT_SECRET=%LOCAL_TEST&@Ub4&9s
 
 S3_LOCAL_STACK=http://localhost:4566
