+++ libs/dal/src/shared/interceptors/invalidate-cache.interceptor.ts
@@ -1,16 +1,15 @@
-import { CacheService } from '../services/cache/cache.service';
-
-const cacheService = new CacheService();
 const USECASE_FUNCTION_NAME = 'execute';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export function Cached(storeKey?: string) {
+export function InvalidateCache(storeKey?: string) {
   return (target: any, key: string, descriptor: any) => {
     const originalMethod = descriptor.value;
     const context = getKeyContext(key, target);
 
     descriptor.value = async function (...args: any[]) {
       const command = args[0];
+
+      // if (command.invalidate !== true) return originalMethod.apply(this, args);
 
       const cacheKey = buildKey(context, command);
 
@@ -19,23 +18,13 @@ export function Cached(storeKey?: string) {
       }
 
       try {
-        const value = await cacheService.get(cacheKey);
-        if (value) {
-          return JSON.parse(value);
-        }
-
-        try {
-          const response = await originalMethod.apply(this, args);
-          await cacheService.set(cacheKey, JSON.stringify(response));
-
-          return response;
-        } catch (err) {
-          // Logger.error(`An error has occured when inserting "key: ${key}", "value: ${response}"`, 'CacheInterceptor');
-          return originalMethod.apply(this, args);
-        }
-      } catch (e) {
+        await this.cacheService.delByPattern(cacheKey);
+      } catch (err) {
+        // Logger.error(`An error has occurred when deleting "key: ${cacheKey}",`, 'InvalidateCache');
         return originalMethod.apply(this, args);
       }
+
+      return originalMethod.apply(this, args);
     };
   };
 
@@ -49,7 +38,7 @@ export function Cached(storeKey?: string) {
 function buildKey(prefix: string, keyConfig: Record<undefined, string>): string {
   let key = prefix;
 
-  key = appendQueryParams(key, keyConfig);
+  key = `${key}*`;
 
   return appendCredentials(key, keyConfig);
 }
@@ -68,24 +57,4 @@ function appendCredentials(key: string, keyConfig: Record<undefined, string>) {
   });
 
   return result;
-}
-
-function appendQueryParams(key: string, keyConfig: any): string {
-  let queryParamsKey = key;
-  const keyElements: Array<string> = ['feedId', 'seen', 'read', 'page', 'query'];
-
-  keyElements.forEach((element) => {
-    if (keyConfig[element] == null) return;
-
-    const elementValue =
-      typeof keyConfig[element] === 'object' ? JSON.stringify(keyConfig[element]) : keyConfig[element];
-
-    const elementKey = `${element}=${elementValue}`;
-
-    if (elementKey) {
-      queryParamsKey += ':' + elementKey;
-    }
-  });
-
-  return queryParamsKey;
 }
