+++ libs/shared/package.json
@@ -20,6 +20,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "dependencies": {
+    "ioredis": "^5.2.4",
     "axios": "^0.26.1",
     "class-transformer": "0.5.1",
     "class-validator": "0.13.2",
