+++ libs/shared/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './http-client';
+export * from './cache';
